refactor: move queryRelationPaged
